==> Final_Project/1_preprocessing.ipynb <==
# Read files
import numpy as np
import pandas as pd

data2020 = pd.read_excel(
    "C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/2021資料.xlsx",
    sheet_name="2020"
)
q12021 = pd.read_excel(
    "C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/2021資料.xlsx",
    sheet_name="Q1不動產"
)
q22021 = pd.read_excel(
    "C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/2021資料.xlsx",
    sheet_name="Q2不動產"
)
q32021 = pd.read_excel(
    "C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/2021資料.xlsx",
    sheet_name="Q3不動產"
)

print("data2020 shape:", data2020.shape)
print("q12021 shape:", q12021.shape)
print("q22021 shape:", q22021.shape)
print("q32021 shape:", q32021.shape)

# Correct「交易年月日」 format in q1~q3
q12021['year'] = q12021['交易年月日'].dt.strftime("%Y").astype(int)
q12021['month'] = q12021['交易年月日'].dt.strftime("%m").astype(int)
q12021['date'] = q12021['交易年月日'].dt.strftime("%d").astype(int)

q22021['year'] = ((q22021['交易年月日'] / 10000).astype(int)) + 1911
q22021['date'] = (q22021['交易年月日'] % 100)
q22021['month'] = (((q22021['交易年月日'] % 10000) - q22021['date']) / 100 ).astype(int)

q32021['year'] = ((q32021['交易年月日'] / 10000).astype(int)) + 1911
q32021['date'] = (q32021['交易年月日'] % 100)
q32021['month'] = (((q32021['交易年月日'] % 10000) - q32021['date']) / 100 ).astype(int)

data2020['year'] = ((data2020['交易年月日'] / 10000).astype(int)) + 1911
data2020['date'] = (data2020['交易年月日'] % 100)
data2020['month'] = (((data2020['交易年月日'] % 10000) - data2020['date']) / 100 ).astype(int)

# Inspect their columns
# Find columns in 2020 data not in 2021 data
spare = []
print("Columns in 2020 not in 2021:")
for c in data2020.columns:
    if c not in q12021.columns:
        spare.append(c)
        print(c)

# Remove these columns
for c in spare:
    data2020 = data2020.drop(c, axis=1)

# Now the columns in two data are the same

# Concatenate altogether
df = pd.concat([data2020, q12021, q22021, q32021])
print("df shape:", df.shape)

# Fill null value in '移轉層次項目'
df['移轉層次項目'] = df['移轉層次項目'].fillna('無')

# Abandon other data with null values
df = df.dropna(axis=0, how = 'any')
print("df shape:", df.shape)

# Construct Housing Age
df['建築完成年月'] = df['建築完成年月'].dt.strftime("%Y").astype(str).str[:4].astype(int)
df['屋齡'] = df['year'] - df['建築完成年月']
print("df shape:", df.shape)

# Other variables preprocessing 
df['有無管理組織'] = df['有無管理組織'].str.contains('有').astype(int)
df['建物現況格局-隔間'] = df['建物現況格局-隔間'].str.contains('有').astype(int)
df = df.drop(df[df['車位數'] == '位'].index)
df['車位數'] = df['車位數'].astype(int)

print("df shape:", df.shape)

# Dummy Variables
dummy_1 = pd.get_dummies(df['建物型態'], prefix='建物型態')
df = pd.concat([dummy_1,df],axis=1)
dummy_2 = pd.get_dummies(df['主要用途'], prefix='主要用途')
df = pd.concat([df,dummy_2],axis=1)
dummy_3 = pd.get_dummies(df['鄉鎮市區'], prefix='鄉鎮市區')
df = pd.concat([df,dummy_3],axis=1)
dummy_4 = pd.get_dummies(df['移轉層次項目'], prefix='移轉層次項目')
df = pd.concat([df,dummy_4],axis=1)
dummy_5 = pd.get_dummies(df["主要建材"], prefix="建材")
df = pd.concat([df,dummy_5],axis=1)
dummy_6 = pd.get_dummies(df["都市土地使用分區"], prefix="使用分區")
df = pd.concat([df,dummy_6],axis=1)

df = df.reset_index()
print("df shape:", df.shape)

# Delete Outliers
totalPrice = df["單價元平方公尺"]
Q1 = np.quantile(totalPrice, 1/4)
Q3 = np.quantile(totalPrice, 3/4)
IQR = Q3 - Q1

# Outliers defined as not in the interval (Q1 - 1.5*IQR, Q3 + 1.5*IQR)
outliers_idx = []
for i in range(totalPrice.shape[0]):
    if (totalPrice.iloc[i] > (Q3 + 1.5*IQR)) or (totalPrice.iloc[i] < (Q1 - 1.5*IQR)):
        outliers_idx.append(i)

for i in outliers_idx:
     df = df.drop(i, axis=0)

print("df shape:", df.shape)
print(df.columns)

# Drop meaningless column
toDrop = ["index", "鄉鎮市區", "交易標的",  "移轉層次項目", "建物型態", "主要用途",
          "主要建材", "都市土地使用分區", "建築完成年月", "year", "month", "date", "總價元", '交易年月日']
# len(toDrop) = 13

for acol in toDrop:
    df = df.drop(acol, axis=1)

# Drop strange data: idx7963
df = df.drop(7963, axis=0)
address = df["土地位置建物門牌"]
df = df.drop("土地位置建物門牌", axis=1)
df = df.apply(pd.to_numeric)  # turn into numerical value
df = pd.concat([df, address], axis=1).reset_index().drop('index', axis=1)

print("df shape:", df.shape)

# Standardization
from sklearn.preprocessing import StandardScaler

num_columns = ['土地移轉總面積平方公尺','土地數','建物數','車位數','移轉層次','總樓層數',
               '建物移轉總面積平方公尺','建物現況格局-房','建物現況格局-廳','建物現況格局-衛',
               '建物現況格局-隔間','屋齡']

scaler = StandardScaler()
numerical_df = scaler.fit_transform(df[num_columns])
numerical_df = pd.DataFrame(numerical_df, columns=num_columns)

# merge into df
df = df.drop(num_columns, axis=1)
df = pd.concat([numerical_df, df], axis=1)
print("df shape:", df.shape)

# Remove feature appears less than 10 times
minN = 10
for col in df.columns[12:-1]:  # dummy variables start at idx.12
#     print("col =", col)
#     print(df_dropped[col].sum())
#     print("-"*80)
    if df[col].sum() < minN:
        df = df.drop(col, axis=1)
print("df shape:", df.shape)

# Inspect all columns
for i in df.columns:
    print(i)

# Set labels
y = df["單價元平方公尺"]
x = df.drop(["單價元平方公尺"], axis=1)

# Split training and test data
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.2)

# Validation data
X_train, X_valid, y_train, y_valid = train_test_split(X_train, y_train, test_size=0.1)
print("X_train:", X_train.shape)
print("X_valid:", X_valid.shape)
print("X_test:", X_test.shape)
print("y_train:", y_train.shape)
print("y_test:", y_test.shape)
print("y_valid:", y_valid.shape)

# X_train.to_csv("C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/DATASET/X_train.csv")
# X_valid.to_csv("C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/DATASET/X_valid.csv")
# X_test.to_csv("C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/DATASET/X_test.csv")
# y_train.to_csv("C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/DATASET/y_train.csv")
# y_test.to_csv("C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/DATASET/y_test.csv")
# y_valid.to_csv("C:/Users/andyi/OneDrive/桌面/統計學習與深度學習/Final Project/data/DATASET/y_valid.csv")
