==> Final_Project/4_b_VGG16_extract_google_streetview_features.ipynb <==
# from google.colab import drive
# drive.mount('/content/drive')

import pandas as pd
vgg_data = pd.read_csv('/content/drive/MyDrive/py_data/data_vgg_final.csv')

df_raw = pd.DataFrame(vgg_data)
df_fulldata = df_raw.drop(columns=['備註1', '備註2', '備註3'])
df_clean = df_fulldata.rename(columns={'地址': 'location', '角度': 'heading', '分數':'score'}).dropna()

df_clean['heading'] = df_clean.heading.astype(int)
df_clean['score'] = df_clean.score.astype(int)
df_clean['param_zip'] = list(zip(df_clean['location'], df_clean['heading']))
df_clean

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.image as mpimg
import io
import time
import requests
import json
from io import BytesIO
from PIL import Image
import numpy as np

# # Set 'inclde_top = False' to extract fc2 (Dense)

import keras
from keras.applications.vgg16 import VGG16

# Set a size which is multiple of 32
# Set 'inclde_top = False' to extract the second last layer of VGG16
SIZE = 224
VGG_model = VGG16(weights='imagenet', include_top=True, input_shape=(SIZE, SIZE, 3))

# Make loaded layers as non-trainable
for layer in VGG_model.layers:
	layer.trainable = False

VGG_model.summary()  #Trainable parameters will be 0

# from keras.applications import vgg16
from keras.preprocessing import image
from keras.applications.vgg16 import preprocess_input
from keras.models import Model
# import numpy as np
# import matplotlib.pyplot as plt

start = 13000
end = 13500
all_data = df_clean['param_zip'].tolist()
data_run = all_data[start:end]
print(len(data_run))

# def my_vgg16(data_run):

# get image from google street static api
pic_base = 'https://maps.googleapis.com/maps/api/streetview?' 
api_key = 'YOUR_API_KEY'

arr_lst = []
print(data_run[0])
print(data_run[-1])

for i in range(len(data_run)):
  current_loc = data_run[i][0]
  current_hd = data_run[i][1]
  
  
  pic_params = {'key':api_key, 'location':current_loc, 'size':"224x224", 'fov':120, 'heading':current_hd} # fov為平視寬窄角度
  pic_response = requests.get(pic_base, params=pic_params)
  im = Image.open(BytesIO(pic_response.content))

  # convert image to array
  image_array = image.img_to_array(im) # keras
  img_arr_lst = image_array.tolist()
  arr_lst.append(img_arr_lst)

print(np.array(arr_lst).shape)

x = np.array(arr_lst)
x = preprocess_input(x)

model_extractfeatures = Model(inputs = VGG_model.input, outputs = VGG_model.get_layer('fc2').output)


fc2_features = model_extractfeatures.predict(x)


print(fc2_features.shape)
np.save('features_fc2_13000_13500', fc2_features)
