==> Final_Project/6_Main.ipynb <==
## Dataset

<li> 實價登錄：x_train_raw
<li>實價登錄 + 機能：x_train_fac
<li>實價登錄 + 機能 + 街景 + 人工標籤：x_train_fac_street
<li>實價登錄 + 機能 + 衛星 + 人工標籤：x_train_fac_satel
<li>實價登錄 + 機能 + 街景 + 衛星 + 人工標籤：x_train_full

import pandas as pd
import numpy as np

# dataset loading
x_train_full = pd.read_csv('C:\\Users\\andyi\\OneDrive\\桌面\\x_train\\x_train_full.csv')
x_train_full = x_train_full.drop(columns=['Unnamed: 0'])
x_valid_full = pd.read_csv('C:\\Users\\andyi\\OneDrive\\桌面\\x_valid\\x_valid_full.csv')
x_valid_full = x_valid_full.drop(columns=['Unnamed: 0'])
x_test_full = pd.read_csv('C:\\Users\\andyi\\OneDrive\\桌面\\x_test\\x_test_full.csv')
x_test_full = x_test_full.drop(columns=['Unnamed: 0'])
ntrain = x_train_full.shape[0]
nvalid = x_valid_full.shape[0]
ntest = x_test_full.shape[0]
X = pd.concat([x_train_full, x_valid_full, x_test_full], axis=0).reset_index(drop=True)

col = X.columns
raw = X[col[:81]]
fac = X[col[81:96]]
street = X[col[96:160]]
satellite = X[col[160:220]]
score = X[col[220]]

# Standardization: fac, street, score
from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
fac = scaler.fit_transform(fac)
street = scaler.fit_transform(street)
score = scaler.fit_transform(np.array(score).reshape(-1,1))

fac = pd.DataFrame(fac, columns=col[81:96])
street = pd.DataFrame(street, columns=[("Street" + str(i)) for i in range(1, 65)])
score = pd.DataFrame(score, columns=[col[220]])
satellite.columns = [("Sat" + str(i)) for i in range(1, 61)]

# Assign to each dataset

def assign(data):
    ALL = pd.concat([d for d in data], axis=1)
    x_train = ALL[:ntrain]
    x_valid = ALL[ntrain:(ntrain+nvalid)]
    x_test = ALL[(ntrain+nvalid):]
    return x_train, x_valid, x_test

# 1. x_train_raw, x_valid_raw, x_test_raw
x_train_raw, x_valid_raw, x_test_raw = assign([raw])

# 2. x_train_fac, x_valid_fac, x_test_fac
x_train_fac, x_valid_fac, x_test_fac = assign([raw, fac])

# 3. x_train_fac_street, x_valid_fac_street, x_test_fac_street
x_train_fac_street, x_valid_fac_street, x_test_fac_street = assign([raw, fac, street, score])

# 4. x_train_fac_satel, x_valid_fac_satel, x_test_fac_satel
x_train_fac_satel, x_valid_fac_satel, x_test_fac_satel = assign([raw, fac, satellite, score])

# 5. x_train_fac_full, x_valid_fac_full, x_test_fac_full
x_train_full, x_valid_full, x_test_full = assign([raw, fac, street, satellite, score])

# y
y_train = pd.read_csv('C:\\Users\\andyi\\OneDrive\\桌面\\y\\y_train.csv')
y_train = y_train.drop(columns=['Unnamed: 0']).to_numpy().ravel()
y_valid = pd.read_csv('C:\\Users\\andyi\\OneDrive\\桌面\\y\\y_valid.csv')
y_valid = y_valid.drop(columns=['Unnamed: 0']).to_numpy().ravel()
y_test = pd.read_csv('C:\\Users\\andyi\\OneDrive\\桌面\\y\\y_test.csv')
y_test = y_test.drop(columns=['Unnamed: 0']).to_numpy().ravel()
y_all = np.concatenate([y_train, y_valid, y_test], axis=0)

## Perform Feature Selection via F-score

from sklearn.feature_selection import SelectKBest, f_regression

def FSelect(dataset, k=100):
    # Concatenate all data
    all_X = pd.concat(dataset, axis=0)
    all_Y = y_all

    # F-test
    selector = SelectKBest(f_regression, k=k)
    all_X_new = selector.fit_transform(all_X, all_Y)
    new_columns = selector.get_feature_names_out()
    all_X_new = pd.DataFrame(all_X_new, columns=new_columns)

    x_train = all_X_new[:ntrain]
    x_valid = all_X_new[ntrain:(ntrain + nvalid)]
    x_test = all_X_new[(ntrain + nvalid):]

    return x_train, x_valid, x_test

# 1. x_train_raw, x_valid_raw, x_test_raw: no need to fselect
# x_train_raw, x_valid_raw, x_test_raw = FSelect([x_train_raw, x_valid_raw, x_test_raw], k=)

# 2. x_train_fac, x_valid_fac, x_test_fac: no need to fselect
# x_train_fac, x_valid_fac, x_test_fac = FSelect([x_train_fac, x_valid_fac, x_test_fac])

# 3. x_train_fac_street, x_valid_fac_street, x_test_fac_street
x_train_fac_street, x_valid_fac_street, x_test_fac_street = FSelect([x_train_fac_street, x_valid_fac_street, x_test_fac_street], k=110)

# 4. x_train_fac_satel, x_valid_fac_satel, x_test_fac_satel
x_train_fac_satel, x_valid_fac_satel, x_test_fac_satel = FSelect([x_train_fac_satel, x_valid_fac_satel, x_test_fac_satel], k=110)

# 5. x_train_fac_full, x_valid_fac_full, x_test_fac_full
x_train_full, x_valid_full, x_test_full = FSelect([x_train_full, x_valid_full, x_test_full], k=150)

print("X_train:", x_train_full.shape)
print("X_test:", x_test_full.shape)
print("X_valid:", x_valid_full.shape)

# Model Define

## Ridge Regression

from sklearn.linear_model import Ridge
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def Ridgereg(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global ridge
    ridge = Ridge()
    col = X_train.columns
    tune_params = {'alpha':np.arange(1, 5, 0.1)}
    gd = GridSearchCV(ridge, tune_params)
    gd.fit(X_valid, y_valid)

    alpha = gd.best_params_["alpha"]
    print("best alpha:", alpha)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    ridge = Ridge(alpha=alpha)
    ridge.fit(X_valid_train, y_valid_train)
    pred = ridge.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    # Find the most important 10 features
    abs_coef_idx = np.flip(np.argsort(np.absolute(ridge.coef_)))
    print("The 10 most important features:")
    for i in range(10):
        if i != 9:
            print(col[abs_coef_idx[i]], end=", ")
        else:
            print(col[abs_coef_idx[i]])

    return RMSE

## Lasso Regression

from sklearn.linear_model import Lasso
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def Lassoreg(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global lasso
    col = X_train.columns
    lasso = Lasso()
    tune_params = {'alpha':np.arange(1, 5, 0.1)}
    gd = GridSearchCV(lasso, tune_params)
    gd.fit(X_valid, y_valid)

    alpha = gd.best_params_["alpha"]
    print("best alpha:", alpha)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    lasso = Lasso(alpha=alpha)
    lasso.fit(X_valid_train, y_valid_train)
    pred = lasso.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    # Find the most important 10 features
    abs_coef_idx = np.flip(np.argsort(np.absolute(lasso.coef_)))
    print("The 10 most important features:")
    for i in range(10):
        if i != 9:
            print(col[abs_coef_idx[i]], end=", ")
        else:
            print(col[abs_coef_idx[i]])
    return RMSE

## KNN

from sklearn.neighbors import KNeighborsRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def KNN(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global knn
    knn = KNeighborsRegressor()
    tune_params = {'n_neighbors':range(1, 31), 'weights':['distance', 'uniform'],
                   'leaf_size':range(1, 20)}
    gd = GridSearchCV(knn, tune_params)
    gd.fit(X_valid, y_valid)

    n_neighbors = gd.best_params_["n_neighbors"]
    print("best n_neighbors:", n_neighbors)
    weights = gd.best_params_["weights"]
    print("best weights:", weights)
    leaf_size = gd.best_params_["leaf_size"]
    print("best leaf_size:", leaf_size)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    knn = KNeighborsRegressor(n_neighbors=n_neighbors, weights=weights, leaf_size=leaf_size)
    knn.fit(X_valid_train, y_valid_train)
    pred = knn.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)
    return RMSE

## Random Forest

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def RandomForest(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global RF
    RF = RandomForestRegressor(random_state=439)
    tune_params = {'n_estimators':range(200, 300, 20)}
    gd = GridSearchCV(RF, tune_params)
    gd.fit(X_valid, y_valid)

    n_estimators = gd.best_params_["n_estimators"]
    print("best n_estimators:", n_estimators)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    RF = RandomForestRegressor(n_estimators=n_estimators, oob_score=True, random_state=439)
    RF.fit(X_valid_train, y_valid_train)
    pred = RF.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    # printing the 10 most important features
    importance = RF.feature_importances_
    idx = np.flip(np.argsort(importance))
    important_10feat = X_train.columns[idx[:10]]
    print("The 10 most important features:")
    for i in range(10):
        if i != 9:
            print(important_10feat[i], end=", ")
        else:
            print(important_10feat[i])
    return RF.oob_score_, RMSE

## Bagging

from sklearn.ensemble import BaggingRegressor, RandomForestRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def Bagging(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global bagging
    bagging = BaggingRegressor(random_state=439)
    tune_params = {'n_estimators':range(200, 300, 20)}
    gd = GridSearchCV(bagging, tune_params)
    gd.fit(X_valid, y_valid)

    n_estimators = gd.best_params_["n_estimators"]
    print("best n_estimators:", n_estimators)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    bagging = BaggingRegressor(n_estimators=n_estimators,
                               oob_score=True,
                               random_state=439)
    bagging.fit(X_valid_train, y_valid_train)
    pred = bagging.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    return bagging.oob_score_, RMSE

## Decision Tree

from sklearn.tree import DecisionTreeRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def DecisionTree(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global DT
    DT = DecisionTreeRegressor(random_state=439)
    tune_params = {'criterion':["squared_error", "friedman_mse", "absolute_error", "poisson"],
                   'max_depth':range(5, 20),
                   'min_samples_split':range(2, 10)}
    gd = GridSearchCV(DT, tune_params, scoring='neg_root_mean_squared_error')
    gd.fit(X_valid, y_valid)

    criterion = gd.best_params_["criterion"]
    max_depth = gd.best_params_["max_depth"]
    min_samples_split = gd.best_params_["min_samples_split"]
    print("best criterion:", criterion)
    print("best max_depth:", max_depth)
    print("best min_samples_split:", min_samples_split)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    DT = DecisionTreeRegressor(criterion=criterion, max_depth=max_depth,
                               min_samples_split=min_samples_split, random_state=439)
    DT.fit(X_valid_train, y_valid_train)
    pred = DT.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    # printing the 10 most important features
    importance = DT.feature_importances_
    idx = np.flip(np.argsort(importance))
    important_10feat = X_train.columns[idx[:10]]
    print("The 10 most important features:")
    for i in range(10):
        if i != 9:
            print(important_10feat[i], end=", ")
        else:
            print(important_10feat[i])

    return RMSE

## Gradient Boosting

from sklearn.ensemble import GradientBoostingRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def GBoosting(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global GBDT
    GBDT = GradientBoostingRegressor(random_state=439)
    tune_params = {'loss':["squared_error", "huber", "absolute_error", "quantile"],
                   'learning_rate':[0.01, 0.1, 0.2, 0.4],
                   'n_estimators':range(120, 200, 20)}
    gd = GridSearchCV(GBDT, tune_params, scoring='neg_root_mean_squared_error')
    gd.fit(X_valid, y_valid)

    loss = gd.best_params_["loss"]
    learning_rate = gd.best_params_["learning_rate"]
    n_estimators = gd.best_params_["n_estimators"]
    print("best loss:", loss)
    print("best learning_rate:", learning_rate)
    print("best n_estimators:", n_estimators)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    GBDT = GradientBoostingRegressor(loss=loss, learning_rate=learning_rate, n_estimators=n_estimators,
                                     random_state=439)
    GBDT.fit(X_valid_train, y_valid_train)
    pred = GBDT.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    # printing the 10 most important features
    importance = GBDT.feature_importances_
    idx = np.flip(np.argsort(importance))
    important_10feat = X_train.columns[idx[:10]]
    print("The 10 most important features:")
    for i in range(10):
        if i != 9:
            print(important_10feat[i], end=", ")
        else:
            print(important_10feat[i])

    return RMSE

## Adaboost

from sklearn.ensemble import AdaBoostRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def AdaBoost(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global ada
    ada = AdaBoostRegressor(random_state=439)
    tune_params = {'loss':["linear", "square", "exponential"],
                   'learning_rate':[0.1, 0.3, 0.5, 0.8, 1, 1.5, 2, 2.5, 3],
                   'n_estimators':range(10, 150, 10)}
    gd = GridSearchCV(ada, tune_params, scoring='neg_root_mean_squared_error')
    gd.fit(X_valid, y_valid)

    loss = gd.best_params_["loss"]
    learning_rate = gd.best_params_["learning_rate"]
    n_estimators = gd.best_params_["n_estimators"]
    print("best loss:", loss)
    print("best learning_rate:", learning_rate)
    print("best n_estimators:", n_estimators)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    ada = AdaBoostRegressor(loss=loss, learning_rate=learning_rate, n_estimators=n_estimators,
                            random_state=439)
    ada.fit(X_valid_train, y_valid_train)
    pred = ada.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    # printing the 10 most important features
    importance = ada.feature_importances_
    idx = np.flip(np.argsort(importance))
    important_10feat = X_train.columns[idx[:10]]
    print("The 10 most important features:")
    for i in range(10):
        if i != 9:
            print(important_10feat[i], end=", ")
        else:
            print(important_10feat[i])

    return RMSE

## Supported Vector Machine Regression

from sklearn.svm import SVR
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def SVM(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global svr
    svr = SVR()
    tune_params = {'kernel':["linear", "poly", "rbf", "sigmoid"],
                   'C':np.arange(0.1, 3, 0.1)}
    gd = GridSearchCV(svr, tune_params, scoring='neg_root_mean_squared_error')
    gd.fit(X_valid, y_valid)

    kernel = gd.best_params_["kernel"]
    C = gd.best_params_["C"]
    print("best kernel:", kernel)
    print("best C:", C)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    svr = SVR(kernel=kernel, C=C)
    svr.fit(X_valid_train, y_valid_train)
    pred = svr.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    return RMSE

## Stochastic Gradient Descent

from sklearn.linear_model import SGDRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def SGD(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global sgd
    sgd = SGDRegressor()
    tune_params = {'loss':["squared_error", "huber", "epsilon_insensitive"],
                   'penalty':["l2", "l1"],
                   'alpha':np.logspace(-6, 1, 8, base=10)}
    gd = GridSearchCV(sgd, tune_params, scoring='neg_root_mean_squared_error')
    gd.fit(X_valid, y_valid)

    loss = gd.best_params_["loss"]
    penalty = gd.best_params_["penalty"]
    alpha = gd.best_params_["alpha"]
    print("best loss:", loss)
    print("best penalty:", penalty)
    print("best alpha:", alpha)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    sgd = SGDRegressor(loss=loss, penalty=penalty, alpha=alpha)
    sgd.fit(X_valid_train, y_valid_train)
    pred = sgd.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    return RMSE

## Multi-Layer Perceptron

from sklearn.neural_network import MLPRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def MLP(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global mlp
    mlp = MLPRegressor(early_stopping=True, hidden_layer_sizes=360)
    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    mlp.fit(X_valid_train, y_valid_train)
    pred = mlp.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)
    return RMSE

## XGBoost

from xgboost import XGBRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def XGB(X_train, y_train, X_valid, y_valid, X_test, y_test):
    global xgb
    xgb = XGBRegressor()
    tune_params = {"n_estimators":range(1, 50)}
    gd = GridSearchCV(xgb, tune_params, scoring='neg_root_mean_squared_error')
    gd.fit(X_valid, y_valid)

    n_estimators = gd.best_params_["n_estimators"]
    print("best n_estimators:", n_estimators)
    print("==========")

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    xgb = XGBRegressor(n_estimators=n_estimators)
    xgb.fit(X_valid_train, y_valid_train)
    pred = xgb.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)

    return RMSE

## Stacking

from sklearn.ensemble import StackingRegressor
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error

def Stacking(X_train, y_train, X_valid, y_valid, X_test, y_test):
    ests = [
            ("Random Forest", RF),
            ("Bagging", bagging),
            ("Gradient Boosting", GBDT),
            ("KNN", knn),
            ("SGD", sgd),
            ("Ridge", ridge),
            ("XGB", xgb)
           ]
    stacking = StackingRegressor(estimators=ests)

    X_valid_train = pd.concat([X_valid, X_train], axis=0)
    y_valid_train = np.concatenate([y_valid, y_train], axis=0)

    stacking.fit(X_valid_train, y_valid_train)

    pred = stacking.predict(X_test)
    RMSE = mean_squared_error(y_test, pred, squared=False)
    return RMSE

# Model Training and Evaluation!

## Dataset

<li> 實價登錄：x_train_raw
<li>實價登錄 + 機能：x_train_fac
<li>實價登錄 + 機能 + 街景 + 人工標籤：x_train_fac_street
<li>實價登錄 + 機能 + 衛星 + 人工標籤：x_train_fac_satel
<li>實價登錄 + 機能 + 街景 + 衛星 + 人工標籤：x_train_full

import warnings
warnings.filterwarnings("ignore")

# Evaluation: Solely Government Data
print("==== Solely Government Data ====\n")

print("||Ridge Regression||")
rmse = Ridgereg(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Ridge Regression | RMSE =", rmse, "\n")

print("||Lasso Regression||")
rmse = Lassoreg(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Lasso Regression | RMSE =", rmse, "\n")

print("||KNN||")
rmse = KNN(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("KNN | RMSE =", rmse, "\n")

print("||Random Forest||")
oob, rmse = RandomForest(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Random Forest | RMSE =", rmse)
print("Random Forest | OOB Score =", oob, "\n")

print("||Bagging||")
oob, rmse = Bagging(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Bagging | RMSE =", rmse)
print("Bagging | OOB Score =", oob, "\n")

print("||Decision Tree||")
rmse = DecisionTree(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Decision Tree | RMSE =", rmse, "\n")

print("||Gradient Boosting||")
rmse = GBoosting(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Gradient Boosting | RMSE =", rmse, "\n")

print("||Ada Boosting||")
rmse = AdaBoost(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Ada Boosting | RMSE =", rmse, "\n")

print("||Support Vector Machine||")
rmse = SVM(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Support Vector Machine | RMSE =", rmse, "\n")

print("||Stochastic Gradient Descent||")
rmse = SGD(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Stochastic Gradient Descent | RMSE =", rmse, "\n")

print("||Multi-Layer Perceptron||")
rmse = MLP(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Multi-Layer Perceptron | RMSE =", rmse, "\n")

print("||XGBoost||")
rmse = XGB(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("XGBoost | RMSE =", rmse, "\n")

print("||Stacking||")
rmse = Stacking(x_train_raw, y_train, x_valid_raw, y_valid, x_test_raw, y_test)
print("Stacking | RMSE =", rmse)

# Evaluation: Government Data with Facility Information
print("==== Government Data with Facility Information ====\n")

print("||Ridge Regression||")
rmse = Ridgereg(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Ridge Regression | RMSE =", rmse, "\n")

print("||Lasso Regression||")
rmse = Lassoreg(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Lasso Regression | RMSE =", rmse, "\n")

print("||KNN||")
rmse = KNN(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("KNN | RMSE =", rmse, "\n")

print("||Random Forest||")
oob, rmse = RandomForest(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Random Forest | RMSE =", rmse)
print("Random Forest | OOB Score =", oob, "\n")

print("||Bagging||")
oob, rmse = Bagging(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Bagging | RMSE =", rmse)
print("Bagging | OOB Score =", oob, "\n")

print("||Decision Tree||")
rmse = DecisionTree(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Decision Tree | RMSE =", rmse, "\n")

print("||Gradient Boosting||")
rmse = GBoosting(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Gradient Boosting | RMSE =", rmse, "\n")

print("||Ada Boosting||")
rmse = AdaBoost(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Ada Boosting | RMSE =", rmse, "\n")

print("||Support Vector Machine||")
rmse = SVM(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Support Vector Machine | RMSE =", rmse, "\n")

print("||Stochastic Gradient Descent||")
rmse = SGD(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Stochastic Gradient Descent | RMSE =", rmse, "\n")

print("||Multi-Layer Perceptron||")
rmse = MLP(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Multi-Layer Perceptron | RMSE =", rmse, "\n")

print("||XGBoost||")
rmse = XGB(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("XGBoost | RMSE =", rmse, "\n")

print("||Stacking||")
rmse = Stacking(x_train_fac, y_train, x_valid_fac, y_valid, x_test_fac, y_test)
print("Stacking | RMSE =", rmse)

# Evaluation: Above data with Street Scene Data
print("==== Above data with Street Scene Data ====\n")


print("||Ridge Regression||")
rmse = Ridgereg(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Ridge Regression | RMSE =", rmse, "\n")

print("||Lasso Regression||")
rmse = Lassoreg(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Lasso Regression | RMSE =", rmse, "\n")

print("||KNN||")
rmse = KNN(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("KNN | RMSE =", rmse, "\n")

print("||Random Forest||")
oob, rmse = RandomForest(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Random Forest | RMSE =", rmse)
print("Random Forest | OOB Score =", oob, "\n")

print("||Bagging||")
oob, rmse = Bagging(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Bagging | RMSE =", rmse)
print("Bagging | OOB Score =", oob, "\n")

print("||Decision Tree||")
rmse = DecisionTree(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Decision Tree | RMSE =", rmse, "\n")

print("||Gradient Boosting||")
rmse = GBoosting(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Gradient Boosting | RMSE =", rmse, "\n")

print("||Ada Boosting||")
rmse = AdaBoost(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Ada Boosting | RMSE =", rmse, "\n")

print("||Support Vector Machine||")
rmse = SVM(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Support Vector Machine | RMSE =", rmse, "\n")

print("||Stochastic Gradient Descent||")
rmse = SGD(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Stochastic Gradient Descent | RMSE =", rmse, "\n")

print("||Multi-Layer Perceptron||")
rmse = MLP(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Multi-Layer Perceptron | RMSE =", rmse, "\n")

print("||XGBoost||")
rmse = XGB(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("XGBoost | RMSE =", rmse, "\n")

print("||Stacking||")
rmse = Stacking(x_train_fac_street, y_train, x_valid_fac_street, y_valid, x_test_fac_street, y_test)
print("Stacking | RMSE =", rmse)

# Evaluation: Above data with Satellite data
print("==== Above data with Satellite data ====\n")

print("||Ridge Regression||")
rmse = Ridgereg(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Ridge Regression | RMSE =", rmse, "\n")

print("||Lasso Regression||")
rmse = Lassoreg(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Lasso Regression | RMSE =", rmse, "\n")

print("||KNN||")
rmse = KNN(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("KNN | RMSE =", rmse, "\n")

print("||Random Forest||")
oob, rmse = RandomForest(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Random Forest | RMSE =", rmse)
print("Random Forest | OOB Score =", oob, "\n")

print("||Bagging||")
oob, rmse = Bagging(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Bagging | RMSE =", rmse)
print("Bagging | OOB Score =", oob, "\n")

print("||Decision Tree||")
rmse = DecisionTree(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Decision Tree | RMSE =", rmse, "\n")

print("||Gradient Boosting||")
rmse = GBoosting(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Gradient Boosting | RMSE =", rmse, "\n")

print("||Ada Boosting||")
rmse = AdaBoost(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Ada Boosting | RMSE =", rmse, "\n")

print("||Support Vector Machine||")
rmse = SVM(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Support Vector Machine | RMSE =", rmse, "\n")

print("||Stochastic Gradient Descent||")
rmse = SGD(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Stochastic Gradient Descent | RMSE =", rmse, "\n")

print("||Multi-Layer Perceptron||")
rmse = MLP(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Multi-Layer Perceptron | RMSE =", rmse, "\n")

print("||XGBoost||")
rmse = XGB(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("XGBoost | RMSE =", rmse, "\n")

print("||Stacking||")
rmse = Stacking(x_train_fac_satel, y_train, x_valid_fac_satel, y_valid, x_test_fac_satel, y_test)
print("Stacking | RMSE =", rmse)

# Evaluation: All data included
print("==== All data included ====\n")

print("||Ridge Regression||")
rmse = Ridgereg(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Ridge Regression | RMSE =", rmse, "\n")

print("||Lasso Regression||")
rmse = Lassoreg(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Lasso Regression | RMSE =", rmse, "\n")

print("||KNN||")
rmse = KNN(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("KNN | RMSE =", rmse, "\n")

print("||Random Forest||")
oob, rmse = RandomForest(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Random Forest | RMSE =", rmse)
print("Random Forest | OOB Score =", oob, "\n")

print("||Bagging||")
oob, rmse = Bagging(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Bagging | RMSE =", rmse)
print("Bagging | OOB Score =", oob, "\n")

print("||Decision Tree||")
rmse = DecisionTree(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Decision Tree | RMSE =", rmse, "\n")

print("||Gradient Boosting||")
rmse = GBoosting(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Gradient Boosting | RMSE =", rmse, "\n")

print("||Ada Boosting||")
rmse = AdaBoost(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Ada Boosting | RMSE =", rmse, "\n")

print("||Support Vector Machine||")
rmse = SVM(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Support Vector Machine | RMSE =", rmse, "\n")

print("||Stochastic Gradient Descent||")
rmse = SGD(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Stochastic Gradient Descent | RMSE =", rmse, "\n")

print("||Multi-Layer Perceptron||")
rmse = MLP(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Multi-Layer Perceptron | RMSE =", rmse, "\n")

print("||XGBoost||")
rmse = XGB(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("XGBoost | RMSE =", rmse, "\n")

print("||Stacking||")
rmse = Stacking(x_train_full, y_train, x_valid_full, y_valid, x_test_full, y_test)
print("Stacking | RMSE =", rmse)
